requirements.txt
numpy
opencv-python
Pillow
scikit-image
rich-cli
questionary